refactoring of GeneticAlgorithmTest
